fix: audio data input was Long changed to float

+++ alignment/utils.py
@@ -30,10 +30,11 @@ def get_parameter_norm(parameters, norm_type = 2):
 
     return total_norm
 
+
 def apply_attention_make_batch(tensor,mask,index,tbtt_step,x_length,y_length):
     max_length = x_length // y_length
     #print(max_length)
-    tensor_pad = torch.LongTensor(tensor.size(0),tensor.size(1), tbtt_step * max(max_length))
+    tensor_pad = torch.FloatTensor(tensor.size(0),tensor.size(1), tbtt_step * max(max_length))
     tensor_pad.zero_()
 
     mask_pad = torch.ones(tensor.size(0), tbtt_step * max(max_length))
